# backend/internal/handlers/query.go
package handlers

import (
	"encoding/json"
	"errors"
	"fmt"
	"io"

	"github.com/gin-gonic/gin"
	"github.com/jiangfire/cornerstone/backend/internal/middleware"
	"github.com/jiangfire/cornerstone/backend/pkg/db"
	"github.com/jiangfire/cornerstone/backend/pkg/dto"
	"github.com/jiangfire/cornerstone/backend/pkg/query"
)

// QueryHandler 查询处理器
type QueryHandler struct {
	executor *query.Executor
}

// NewQueryHandler 创建查询处理器
func NewQueryHandler() *QueryHandler {
	return &QueryHandler{
		executor: query.NewExecutor(db.DB()),
	}
}

// Query 统一查询接口
// POST /api/query
// GET /api/query?q={json_string}
//
// @Summary      Execute a query
// @Description  Execute a full Query DSL request. Supports from, select, where, order_by, limit, offset, group_by, having, aggregates, join, and union.
// @Tags         query
// @Accept       json
// @Produce      json
// @Security     ApiKeyAuth
// @Param        body  body  object  true  "Query DSL body"  example({"from":"records","select":["id","data"],"where":{"and":[{"field":"table_id","op":"eq","value":"tbl-1"}]},"orderBy":[{"field":"created_at","dir":"desc"}],"page":1,"size":20})
// @Success      200  {object}  map[string]any  "{"code":0,"data":{"data":[...],"total":0,"page":1,"size":20,"has_more":false}}"
// @Failure      400  {object}  map[string]any
// @Failure      403  {object}  map[string]any
// @Router       /query [post]
func (h *QueryHandler) Query(c *gin.Context) {
	userID := middleware.GetTokenID(c)

	var req query.QueryRequest

	// 尝试从 POST body 解析
	if c.Request.Method == "POST" && c.Request.Body != nil {
		body, err := io.ReadAll(c.Request.Body)
		if err == nil && len(body) > 0 {
			if err := json.Unmarshal(body, &req); err != nil {
				dto.BadRequest(c, "请求格式错误: "+err.Error())
				return
			}
		}
	}

	// 如果 body 为空，尝试从 URL 参数解析
	if req.From == "" && req.Table == "" {
		q := c.Query("q")
		if q == "" {
			dto.BadRequest(c, "缺少查询参数")
			return
		}

		if err := json.Unmarshal([]byte(q), &req); err != nil {
			dto.BadRequest(c, "查询格式错误: "+err.Error())
			return
		}
	}

	// 执行查询
	result, err := h.executor.Execute(c.Request.Context(), &req, userID)
	if err != nil {
		// 区分权限错误和其他错误
		if isPermissionError(err) {
			dto.Forbidden(c, err.Error())
			return
		}
		dto.BadRequest(c, err.Error())
		return
	}

	dto.Success(c, result)
}

// QueryExplain 查询解释接口（返回生成的 SQL，用于调试）
// POST /api/query/explain
//
// @Summary      Explain a query
// @Description  Returns the generated SQL and parameters for a query without executing it. Useful for debugging.
// @Tags         query
// @Accept       json
// @Produce      json
// @Security     ApiKeyAuth
// @Param        body  body  object  true  "Query DSL body"  example({"from":"records","select":["id"],"where":{"and":[{"field":"table_id","op":"eq","value":"tbl-1"}]}})
// @Success      200  {object}  map[string]any  "{"code":0,"data":{"sql":"SELECT ...","params":[...]}}"
// @Failure      400  {object}  map[string]any
// @Failure      403  {object}  map[string]any
// @Router       /query/explain [post]
func (h *QueryHandler) QueryExplain(c *gin.Context) {
	userID := middleware.GetTokenID(c)

	var req query.QueryRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		if !errors.Is(err, io.EOF) {
			dto.BadRequest(c, "请求格式错误: "+err.Error())
			return
		}
		// 尝试从 URL 参数解析
		q := c.Query("q")
		if q == "" {
			dto.BadRequest(c, "缺少查询参数")
			return
		}
		if err := json.Unmarshal([]byte(q), &req); err != nil {
			dto.BadRequest(c, "查询格式错误: "+err.Error())
			return
		}
	}

	// 在权限过滤后的上下文中生成 SQL
	sqlQuery, err := h.executor.ExplainAuthorized(c.Request.Context(), &req, userID)
	if err != nil {
		if isPermissionError(err) {
			dto.Forbidden(c, err.Error())
			return
		}
		dto.BadRequest(c, err.Error())
		return
	}

	dto.Success(c, gin.H{
		"sql":    sqlQuery.SQL,
		"params": sqlQuery.Params,
	})
}

// QueryValidate 查询验证接口（验证查询权限，不执行）
// POST /api/query/validate
//
// @Summary      Validate a query
// @Description  Validate a query DSL and check access permissions without executing it.
// @Tags         query
// @Accept       json
// @Produce      json
// @Security     ApiKeyAuth
// @Param        body  body  object  true  "Query DSL body"  example({"from":"records","select":["id"]})
// @Success      200  {object}  map[string]any  "{"code":0,"message":"查询验证通过"}"
// @Failure      400  {object}  map[string]any
// @Failure      403  {object}  map[string]any
// @Router       /query/validate [post]
func (h *QueryHandler) QueryValidate(c *gin.Context) {
	userID := middleware.GetTokenID(c)

	var req query.QueryRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		if !errors.Is(err, io.EOF) {
			dto.BadRequest(c, "请求格式错误: "+err.Error())
			return
		}
		// 尝试从 URL 参数解析
		q := c.Query("q")
		if q == "" {
			dto.BadRequest(c, "缺少查询参数")
			return
		}
		if err := json.Unmarshal([]byte(q), &req); err != nil {
			dto.BadRequest(c, "查询格式错误: "+err.Error())
			return
		}
	}

	// 验证查询
	if err := h.executor.Validate(c.Request.Context(), &req, userID); err != nil {
		dto.Forbidden(c, err.Error())
		return
	}

	dto.SuccessWithMessage(c, "查询验证通过", nil)
}

// BatchQuery 批量查询接口
// POST /api/query/batch
//
// @Summary      Batch query
// @Description  Execute multiple named queries in a single request.
// @Tags         query
// @Accept       json
// @Produce      json
// @Security     ApiKeyAuth
// @Param        body  body  object  true  "Batch query request"  example({"queries":{"q1":{"from":"records","select":["id"]},"q2":{"from":"tables","select":["id","name"]}}})
// @Success      200  {object}  map[string]any  "{"code":0,"data":{"results":{"q1":{...},"q2":{...}}}}"
// @Failure      400  {object}  map[string]any
// @Failure      403  {object}  map[string]any
// @Router       /query/batch [post]
func (h *QueryHandler) BatchQuery(c *gin.Context) {
	userID := middleware.GetTokenID(c)

	var req query.BatchQueryRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		dto.BadRequest(c, "请求格式错误: "+err.Error())
		return
	}

	// 执行批量查询
	result, err := h.executor.ExecuteBatch(c.Request.Context(), &req, userID)
	if err != nil {
		if isPermissionError(err) {
			dto.Forbidden(c, err.Error())
			return
		}
		dto.BadRequest(c, err.Error())
		return
	}

	dto.Success(c, result)
}

// ListTables 获取可访问的表列表
// GET /api/query/tables
//
// @Summary      List queryable tables
// @Description  Returns all tables the authenticated token can query.
// @Tags         query
// @Accept       json
// @Produce      json
// @Security     ApiKeyAuth
// @Success      200  {object}  map[string]any  "{"code":0,"data":{"tables":[...]}}"
// @Failure      500  {object}  map[string]any
// @Router       /query/tables [get]
func (h *QueryHandler) ListTables(c *gin.Context) {
	userID := middleware.GetTokenID(c)

	tables, err := h.executor.GetValidator().GetAllowedTables(c.Request.Context(), userID)
	if err != nil {
		dto.InternalServerError(c, "获取表列表失败: "+err.Error())
		return
	}

	dto.Success(c, gin.H{
		"tables": tables,
	})
}

// GetTableSchema 获取表结构信息
// GET /api/query/schema/:table
//
// @Summary      Get table schema for query
// @Description  Returns the allowed fields for a queryable table.
// @Tags         query
// @Accept       json
// @Produce      json
// @Security     ApiKeyAuth
// @Param        table  path  string  true  "Table name (records, tables, databases, fields, files, tokens)"
// @Success      200  {object}  map[string]any  "{"code":0,"data":{"table":"records","fields":["id","table_id",...]}}"
// @Failure      403  {object}  map[string]any
// @Router       /query/schema/{table} [get]
func (h *QueryHandler) GetTableSchema(c *gin.Context) {
	userID := middleware.GetTokenID(c)
	table := c.Param("table")

	validator := h.executor.GetValidator()

	// 检查表是否允许访问
	if err := validator.CheckTableAccess(c.Request.Context(), userID, table); err != nil {
		dto.Forbidden(c, err.Error())
		return
	}

	fields := query.DefaultAllowedTables.GetAllowedFields(table)

	dto.Success(c, gin.H{
		"table":  table,
		"fields": fields,
	})
}

// SimplifiedQuery 简化查询接口（URL 参数形式）
// GET /api/query/simple?table=records&filter={}&sort=-created_at&page=1&size=20
//
// @Summary      Simplified query
// @Description  Query records using simple URL parameters instead of full DSL.
// @Tags         query
// @Accept       json
// @Produce      json
// @Security     ApiKeyAuth
// @Param        table   query  string  true   "Table name"
// @Param        filter  query  string  false  "JSON filter object"
// @Param        sort    query  string  false  "Sort expression (prefix with - for desc)"  default(-created_at)
// @Param        page    query  int     false  "Page number"  default(1)
// @Param        size    query  int     false  "Page size (max 1000)"  default(20)
// @Success      200  {object}  map[string]any
// @Failure      400  {object}  map[string]any
// @Failure      403  {object}  map[string]any
// @Router       /query/simple [get]
func (h *QueryHandler) SimplifiedQuery(c *gin.Context) {
	userID := middleware.GetTokenID(c)

	table := c.Query("table")
	if table == "" {
		dto.BadRequest(c, "必须指定表名")
		return
	}

	// 解析 filter
	var filter map[string]interface{}
	filterStr := c.Query("filter")
	if filterStr != "" {
		if err := json.Unmarshal([]byte(filterStr), &filter); err != nil {
			dto.BadRequest(c, "filter 格式错误: "+err.Error())
			return
		}
	}

	// 解析其他参数
	sort := c.DefaultQuery("sort", "-created_at")
	page := 1
	size := 20

	if p := c.Query("page"); p != "" {
		if v, err := parseInt(p); err == nil && v > 0 {
			page = v
		}
	}

	if s := c.Query("size"); s != "" {
		if v, err := parseInt(s); err == nil && v > 0 && v <= 1000 {
			size = v
		}
	}

	// 执行简化查询
	result, err := h.executor.SimplifiedQuery(c.Request.Context(), table, filter, sort, page, size, userID)
	if err != nil {
		if isPermissionError(err) {
			dto.Forbidden(c, err.Error())
			return
		}
		dto.BadRequest(c, err.Error())
		return
	}

	dto.Success(c, result)
}

// isPermissionError 判断是否为权限错误
func isPermissionError(err error) bool {
	if err == nil {
		return false
	}
	msg := err.Error()
	permissionKeywords := []string{
		"权限", "无权", "拒绝", "denied", "forbidden",
		"不能访问", "不允许", "未授权", "unauthorized",
	}
	for _, keyword := range permissionKeywords {
		if containsString(msg, keyword) {
			return true
		}
	}
	return false
}

// containsString 检查字符串是否包含子串
func containsString(s, substr string) bool {
	return len(s) >= len(substr) && (s == substr || s != "" && containsSubstring(s, substr))
}

// containsSubstring 简单子串检查
func containsSubstring(s, substr string) bool {
	for i := 0; i <= len(s)-len(substr); i++ {
		if s[i:i+len(substr)] == substr {
			return true
		}
	}
	return false
}

// parseInt 解析整数
func parseInt(s string) (int, error) {
	var n int
	if err := json.Unmarshal([]byte(s), &n); err != nil {
		// 尝试手动解析
		n = 0
		for _, ch := range s {
			if ch < '0' || ch > '9' {
				return 0, fmt.Errorf("invalid number: %s", s)
			}
			n = n*10 + int(ch-'0')
		}
	}
	return n, nil
}